update debian control file (#41)
This updates the debian control file to link
to the github rather than point to some broken
google-internal links.

The rustc and cargo version bumps are not really
correct since the MSRV is 1.74. Debian testing
does not seem to have that version yet though,
so 1.70 was the best I could do. The build at
least works if you install 1.74 via rustup
and then run a debian build.