link to wiki from README
This patch adds a link to the wiki from the README so
that people will be able to find it. Shpool is (hopefully)
a tool that will be able to fit into different people's
workflows in lots of different ways, and the wiki will
be a good place for folks to share their tricks. We
already have one such tip in there.

Closes #49