fix: protocol forward compat
This patch switches us from using bincode, which has
pretty terrible forwards compatibility properties
to using msgpack (I think I did a switch in the
previous direction in the past). bincode is faster,
but we don't actually care about encoding speed
since we are just encoding some tiny baby little
headers at the start of a connection. This
change also adds a bunch of #[serde(default)]
annotations, which should take care of the second
half of forwards compatibilty.

This means one last protocol version break before
hopefully being able to evolve the protocol in a
much more civilized manner.

I realized I was going to have to add a protocol
field for #47, so this should be done first.

Fixes #85