feat: move protocol types to their own crate
This patch copies all the types from protocol.rs
(plus tty::Size) into a new shpool-protocol
crate. All the methods hanging off of them
are stripped out, since this crate is meant
to kept as purely a place for types to live.

Once this crate is published, the next step is
to move all the methods to Ext traits and
delete the old types, then we can implement
some logic to do version negotiation as
described by #88