fix: migrate to new shpool-protocol crate
This patch migrates us from the old protocol
module to the new protocol crate. The logic
remains in the protocol module in the form
of the client struct and the new ChunkExt
extension trait since the methods don't
belong in the protocol description itself.

This does not yet add the version negotiation
logic that we will need before our next release,
but it does get us one step closer to resolving
issue #88.