fix: exit success for graceful detaches
This patch fixes the exit code for the attach
process so that it is 0 when shpool has been
told to gracefully detach, either via the
`shpool detach` subcommand or via the detach
keybinding. It is more semantically correct
to exit 0 in these cases.

Fixes #130

BREAKING CHANGE: technically could impact some scripts