fix: suppress prompt sentinels from history
This prefixes the sentinel commands we use to
bracket the prompt setup code with a space in order to
prevent them from cluttering up the history.
It might not work for all shells, but hopefully
it will improve things a little.

This was brought up in #120 (though it isn't the
main problem from that bug).