fix: deadlock when shell->client thread stops
This patch fixes an issue where we would lock up the
whole daemon if the shell->client thread stopped
answering. Eventually we would try to talk to
the stuck thread with the global shells lock,
which pretty much makes all endpoints hang.

Adding timeouts to all our RPCs to the thread
should at least prevent the daemon from completely
locking up, but the issue causing this to happen
in the first place is still there. We might just
need to clean up the session in question, but it
could also be some other timing issue.