fix: reconnect hangup due to long session restore (#199)
This patch fixes an issue where `shpool attach` would immediately
hang up while trying to reconnect to a session containing a large
restore buffer. The problem was that the heartbeat thread would
time out requesting a heartbeat from the shell->client thread
because it was busy cooking up the giant restore buffer. This
would lead all the other client threads to exit because they
think the socket connection is unhealthy despite it being
completely fine. The fix is simply to `continue` in the face
of a send timeout in the heartbeat thread rather than tearing
everything down.