fix: hang when starting with nu shell (#272)
This patch fixes a bug where shpool would hang when trying to
start a nu shell by default. This has to do with our startup
sentinal injection. A workaround exists where you can just
blank out the prompt prefix, but we should try to automatically
do that much when working with nu.

I should add a note to the troubleshooting wiki explaining this
so that nu shell users don't get confused as to why shpool is
not injecting a prompt automatically for them.

Fixes #269